@@ -20,8 +20,8 @@
     "@evolu/common": "^5.4.8",
     "@evolu/common-web": "^8.2.4",
     "@fontsource-variable/outfit": "^5.1.0",
-    "@fortawesome/free-brands-svg-icons": "^6.7.1",
-    "@fortawesome/free-solid-svg-icons": "^6.7.1",
+    "@fortawesome/free-brands-svg-icons": "^6.7.2",
+    "@fortawesome/free-solid-svg-icons": "^6.7.2",
     "@tauri-apps/api": "^2.1.1",
     "@tauri-apps/plugin-dialog": "^2.2.0",
     "@tauri-apps/plugin-fs": "^2.2.0",
@@ -43,7 +43,7 @@
     "@eslint/js": "^9.17.0",
     "@faker-js/faker": "^9.3.0",
     "@sveltejs/adapter-static": "^3.0.6",
-    "@sveltejs/kit": "^2.11.1",
+    "@sveltejs/kit": "^2.13.0",
     "@sveltejs/vite-plugin-svelte": "^4.0.1",
     "@tailwindcss/typography": "^0.5.15",
     "@tauri-apps/cli": "^2.1.0",
@@ -53,17 +53,17 @@
     "eslint": "^9.17.0",
     "eslint-config-prettier": "^9.1.0",
     "eslint-plugin-svelte": "^2.46.1",
-    "globals": "^15.13.0",
+    "globals": "^15.14.0",
     "postcss": "^8.4.49",
     "prettier": "^3.4.2",
     "prettier-plugin-svelte": "^3.3.2",
     "prettier-plugin-tailwindcss": "^0.6.9",
-    "svelte": "^5.14.0",
+    "svelte": "^5.15.0",
     "svelte-check": "^4.1.1",
-    "tailwindcss": "^3.4.16",
+    "tailwindcss": "^3.4.17",
     "tslib": "^2.8.1",
     "typescript": "^5.7.2",
-    "typescript-eslint": "^8.18.0",
+    "typescript-eslint": "^8.18.1",
     "vite": "^5.4.11",
     "vitest": "^2.1.8"
   }
