@@ -1,8 +1,1 @@
-export const load = async (event) => {
-  const { storage } = await event.parent();
-  const characters = await storage.getAllCharacters();
-
-  return {
-    characters,
-  };
-};
+export const load = async (event) => {};
